requirements.txt
pytest
pytest-instafail
pytest-mock
python-dateutil
selenium
six
